requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
alembic==1.16.2
    # via astramind (pyproject.toml)
amqp==5.3.1
    # via kombu
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
    #   starlette
bcrypt==4.3.0
    # via passlib
billiard==4.2.1
    # via celery
black==25.1.0
    # via astramind (pyproject.toml)
cachetools==5.5.2
    # via google-auth
celery==5.5.3
    # via astramind (pyproject.toml)
certifi==2025.6.15
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   black
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   uvicorn
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
coloredlogs==15.0.1
    # via onnxruntime
deprecated==1.2.18
    # via limits
distro==1.9.0
    # via openai
fastapi==0.115.13
    # via astramind (pyproject.toml)
fastembed==0.7.1
    # via qdrant-client
filelock==3.18.0
    # via huggingface-hub
flatbuffers==25.2.10
    # via onnxruntime
fsspec==2025.5.1
    # via huggingface-hub
google-api-core==2.25.1
    # via
    #   google-cloud-core
    #   google-cloud-documentai
    #   google-cloud-storage
google-auth==2.40.3
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-documentai
    #   google-cloud-storage
google-cloud-core==2.4.3
    # via google-cloud-storage
google-cloud-documentai==3.5.0
    # via astramind (pyproject.toml)
google-cloud-storage==3.1.1
    # via astramind (pyproject.toml)
google-crc32c==1.7.1
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via google-cloud-storage
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.73.0
    # via
    #   google-api-core
    #   grpcio-status
    #   qdrant-client
grpcio-status==1.73.0
    # via google-api-core
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
h2==4.2.0
    # via httpx
hf-xet==1.1.5
    # via huggingface-hub
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   openai
    #   qdrant-client
huggingface-hub==0.33.0
    # via
    #   fastembed
    #   tokenizers
humanfriendly==10.0
    # via coloredlogs
hyperframe==6.1.0
    # via h2
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
isort==6.0.1
    # via astramind (pyproject.toml)
itsdangerous==2.2.0
    # via astramind (pyproject.toml)
jinja2==3.1.6
    # via astramind (pyproject.toml)
jiter==0.10.0
    # via openai
kombu==5.5.4
    # via celery
limits==5.4.0
    # via slowapi
loguru==0.7.3
    # via fastembed
mako==1.3.10
    # via alembic
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
mmh3==5.1.0
    # via fastembed
mpmath==1.3.0
    # via sympy
mypy-extensions==1.1.0
    # via black
numpy==2.3.1
    # via
    #   fastembed
    #   onnxruntime
    #   qdrant-client
onnxruntime==1.22.0
    # via fastembed
openai==1.91.0
    # via astramind (pyproject.toml)
packaging==25.0
    # via
    #   black
    #   huggingface-hub
    #   kombu
    #   limits
    #   onnxruntime
passlib==1.7.4
    # via astramind (pyproject.toml)
pathspec==0.12.1
    # via black
pillow==11.2.1
    # via fastembed
platformdirs==4.3.8
    # via black
portalocker==2.10.1
    # via qdrant-client
prompt-toolkit==3.0.51
    # via click-repl
proto-plus==1.26.1
    # via
    #   google-api-core
    #   google-cloud-documentai
protobuf==6.31.1
    # via
    #   google-api-core
    #   google-cloud-documentai
    #   googleapis-common-protos
    #   grpcio-status
    #   onnxruntime
    #   proto-plus
    #   qdrant-client
psycopg2-binary==2.9.10
    # via astramind (pyproject.toml)
py-rust-stemmers==0.1.5
    # via fastembed
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.7
    # via
    #   astramind (pyproject.toml)
    #   fastapi
    #   openai
    #   qdrant-client
pydantic-core==2.33.2
    # via pydantic
python-dateutil==2.9.0.post0
    # via celery
python-dotenv==1.1.1
    # via astramind (pyproject.toml)
python-multipart==0.0.20
    # via astramind (pyproject.toml)
pyyaml==6.0.2
    # via huggingface-hub
qdrant-client==1.14.3
    # via astramind (pyproject.toml)
redis==6.2.0
    # via astramind (pyproject.toml)
requests==2.32.4
    # via
    #   fastembed
    #   google-api-core
    #   google-cloud-storage
    #   huggingface-hub
rsa==4.9.1
    # via google-auth
six==1.17.0
    # via python-dateutil
slowapi==0.1.9
    # via astramind (pyproject.toml)
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlalchemy==2.0.41
    # via
    #   astramind (pyproject.toml)
    #   alembic
starlette==0.46.2
    # via fastapi
sympy==1.14.0
    # via onnxruntime
tokenizers==0.21.2
    # via fastembed
tqdm==4.67.1
    # via
    #   fastembed
    #   huggingface-hub
    #   openai
typing-extensions==4.14.0
    # via
    #   alembic
    #   fastapi
    #   huggingface-hub
    #   limits
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via kombu
urllib3==2.5.0
    # via
    #   qdrant-client
    #   requests
uvicorn==0.34.3
    # via astramind (pyproject.toml)
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
watchdog==6.0.0
    # via astramind (pyproject.toml)
wcwidth==0.2.13
    # via prompt-toolkit
wrapt==1.17.2
    # via deprecated
